add packadges & changing the initialization procedure for numeric buttons
