+++ Calculator_lesson1/app/src/main/java/com/example/calculator_lesson1/Domain/Calculatable.java
@@ -1,4 +1,4 @@
-package com.example.calculator_lesson1;
+package com.example.calculator_lesson1.Domain;
 
 public interface Calculatable {
